requirements.txt
chdp==4.0.5
dbjson==1.0.2
kiki_random==1.0.1
asyncio==3.4.3
discord.py==1.3.4